Revert " Break the code to classes (#241)"
This reverts commit c49a5a0dac6696986da5bb8d506ed1ebcca7bcca.
